Move ports command from @config ports to @system ports
The previous commit upgraded the existing @config ports stub. Per
discussion, the right home for this is the top-level @system
namespace alongside the other diagnostic commands (@system session,
@system history, @system env, etc.) since it's a system-wide
diagnostic, not a configuration toggle.

  - Extract the handler into its own file
    system/handlers/portsCommandHandler.ts exporting
    PortsCommandHandler (the prior commit's class).
  - Register it under systemHandlers in systemAgent.ts.
  - Remove the now-empty 'ports' entry from getConfigCommandHandlers.
  - Update agentServer/protocol/README.md to mention @system ports
    (not @config ports).

Build clean for agent-dispatcher; tests still 681/681.
